Initial Update
These scripts find the ongoing attack and alert the user with attackers IP. With the pcap files captured during attack, we can geo-locate the attackers IP address. And also generate XML file which has attackers location which when uploaded in google map provide us the map view of attackers location.